move some data classes out of TransactionReviewViewModel
