+++ app/src/main/java/com/babylon/wallet/android/presentation/transaction/analysis/FeesResolver.kt
@@ -1,9 +1,10 @@
 package com.babylon.wallet.android.presentation.transaction.analysis
 
 import com.babylon.wallet.android.data.transaction.NotaryAndSigners
+import com.babylon.wallet.android.domain.model.GuaranteeAssertion
 import com.babylon.wallet.android.presentation.transaction.PreviewType
 import com.babylon.wallet.android.presentation.transaction.fees.TransactionFees
-import com.babylon.wallet.android.presentation.transaction.guaranteesCount
+import com.babylon.wallet.android.presentation.transaction.model.AccountWithTransferableResources
 import com.radixdlt.sargon.ExecutionSummary
 import com.radixdlt.sargon.extensions.compareTo
 import com.radixdlt.sargon.extensions.toDecimal192
@@ -34,3 +35,9 @@ object FeesResolver {
         }
     }
 }
+
+private fun List<AccountWithTransferableResources>.guaranteesCount(): Int = map { accountWithTransferableResources ->
+    accountWithTransferableResources.resources.filter { transferable ->
+        transferable.guaranteeAssertion is GuaranteeAssertion.ForAmount
+    }
+}.flatten().count()
